blocks: ask for the prompt ids of a RANGE, not of the file's tail
`covers` came back empty on every historical block — 72 of 72 on a real 20 MB
transcript — and `covers[].prompt_id` is what Atlas joins its display on.

The emitter took its ids from resolve.RecentPromptIDs, which reads a bounded
16 MB FILE TAIL. That is right for its own consumer (a model's context window,
which is about the present) and wrong here: the emitter drains blocks
CHRONOLOGICALLY from a persisted cursor, so on a transcript over the tail every
block whose prompts sit below `size - 16 MB` publishes an empty mapping.

So a SIBLING capability rather than a widening of that one — resolve.RangeIDReader
/ PromptIDsInRange, ids only, never text, exactly as RecentPromptIDs is a sibling
of RecentPrompts. Transcripts are chronological and append-only, so the start of a
range is binary-searchable (~11 probes on a 90 MB file); the forward scan from
there stops as soon as a timestamp passes the range, keeping the cost proportional
to the SPAN rather than the file. Measured on an 86 MB live transcript: three
head-of-file hours answered in 3.3/3.9/2.8 ms, ground the tail scan cannot see at
all.

The last human prompt at or before the range LEADS the answer, via a doubling
backward search when the bracket holds none. That is the long-running-episode
case `covers` exists for: an episode that began before the block still covers the
block's opening, and without it the block reads as unattended work.

Nothing assumes strict monotonicity (9 in 9,937 real turns precede the line
before them): both edges are bracketed by five minutes, every membership decision
is made against the record's own timestamp, and the answer is sorted ascending by
instant. Malformed and timestamp-less lines are skipped, never fatal.

The emitter makes ONE range call per sweep over [cursor, now] — the span being
drained. It cannot be narrowed to the exact block edges because the ids ride the
same request that asks for the blocks.

idTailBytes becomes a var so the regression test can shrink it: the failure is a
property of the bound relative to the range asked for, not of any absolute size,
so a 400-byte fixture expresses it exactly where a generated 16 MB one would
express it incidentally and cost every `go test ./...` 16 MB of writing. Verified
the test fails against the old tail-only behaviour.

+++ internal/agent/resolve/recentids_test.go
@@ -121,7 +121,7 @@ func TestRecentPromptIDsCostOnAFullBudgetTail(t *testing.T) {
 		t.Fatal(err)
 	}
 	blob := strings.Repeat("x", 64*1024)
-	written := 0
+	written := int64(0)
 	prompts := 0
 	for written < idTailBytes+(2<<20) {
 		// One human prompt per ~1.3 MB of agent output, the density measured on
@@ -132,7 +132,7 @@ func TestRecentPromptIDsCostOnAFullBudgetTail(t *testing.T) {
 			"message":  map[string]any{"role": "user", "content": "do the thing"},
 		})
 		n, _ := f.Write(append(line, '\n'))
-		written += n
+		written += int64(n)
 		prompts++
 		for i := 0; i < 20; i++ {
 			tr, _ := json.Marshal(map[string]any{
@@ -142,7 +142,7 @@ func TestRecentPromptIDsCostOnAFullBudgetTail(t *testing.T) {
 					"content": []any{map[string]any{"type": "tool_result", "content": blob}}},
 			})
 			n, _ := f.Write(append(tr, '\n'))
-			written += n
+			written += int64(n)
 		}
 	}
 	f.Close()
